cleanup: 6 safety fixes from clang-tidy worklist (#766 / #770)
Fixes the medium-priority cluster from the lint worklist that's surfaced by
the .clang-tidy baseline in #770. Each is a small, isolated safety guard:

  - src/utils/cas/html.c:89 -- fopen(template,"r") result wasn't NULL-
    checked before fgetc() dereferenced it; a missing template file
    crashed the CGI helper.  Add the NULL-check + perror+exit to match
    the existing calloc/fstat error-handling shape in this file.

  - src/PartFile.cpp:3341, 3420 -- loop variable promoted from uint16
    to uint32 to match `partCount` (declared uint32 at line 3193).
    GetPartCount() currently returns uint16 so today's behaviour is
    unchanged, but a future widening wouldn't silently truncate the
    iteration above 65535.  Fires bugprone-too-small-loop-variable.

  - src/utils/aLinkCreator/src/ed2khash.cpp:114 -- user-cancel path
    (progress hook returning false) freed `buf` but leaked the
    realloc()-grown `tmpCharHash`.  Hash of every mid-run cancelled
    file leaked the cumulative parthash buffer.  Free both.  Fires
    clang-analyzer-unix.Malloc.

  - src/kademlia/kademlia/Search.cpp:241 -- `(time_t)(uint32 + 3) >
    time(NULL)` did the addition in uint32 before the cast, so the
    comparison would reorder near the 2106 32-bit time wraparound.
    Cast first, add second.  Eighty years out, but cheap to write
    correctly.  Fires bugprone-misplaced-widening-cast.

  - src/ED2KLinkParser.cpp:107 -- `string(getenv("HOME"))` is UB if
    HOME is unset (rare but possible).  Mirror the existing
    macOS-branch pattern further up the same function:
    `string(home ? home : "")`.  Fires
    clang-analyzer-cplusplus.StringChecker.

  - src/BitVector.h:144 -- SetBuffer() ran memcpy(m_vector, src,
    m_bytes) unconditionally.  After clear(), m_vector is NULL and
    m_bytes is 0, so the call reduces to memcpy(NULL, src, 0), which
    is C-standard UB even though real libcs no-op it.  Guard on
    m_bytes (matches SetAllTrue() above).  Fires
    clang-analyzer-core.NonNullParamChecker.

All builds clean across amule, amuled, amulegui, cas, alc on macOS.
Re-run of clang-tidy with the #770 baseline confirms each of the
above sites no longer fires; no new warnings introduced.

+++ src/PartFile.cpp
@@ -3338,7 +3338,7 @@ void CPartFile::FlushBuffer(bool fromAICHRecoveryDataAvailable)
 		&& !m_hashsetneeded
 		&& m_pendingHashes == 0
 		&& !m_gaplist.IsComplete()) {
-		for (uint16 partNumber = 0; partNumber < partCount; ++partNumber) {
+		for (uint32 partNumber = 0; partNumber < partCount; ++partNumber) {
 			if (!m_aChangedPart[partNumber]) {
 				continue;
 			}
@@ -3416,8 +3416,8 @@ void CPartFile::FlushBuffer(bool fromAICHRecoveryDataAvailable)
 		// download.  The quiescent guard above (1 s no receives)
 		// guarantees writes have drained before we enqueue, so the
 		// contention case can't arise.
-		uint16 enqueued = 0;
-		for (uint16 partNumber = 0; partNumber < partCount; ++partNumber) {
+		uint32 enqueued = 0;
+		for (uint32 partNumber = 0; partNumber < partCount; ++partNumber) {
 			if (!m_aChangedPart[partNumber]) {
 				continue;
 			}
